requirements.txt
requests
webdriver_manager
selenium
argparse
fake-useragent
dnspython3
